Lift contents of integration_tests/flowmachine_server_tests/conftest.py into integration_tests/conftest.py so that the correct flowmachine connection is used in all tests (and cleanup happens automatically, too)
