+++ integration_tests/tests/flowmachine_tests/conftest.py
@@ -5,20 +5,6 @@ from approvaltests.reporters.generic_diff_reporter_factory import (
     GenericDiffReporterFactory,
 )
 
-import flowmachine
-from flowmachine.core import Query
-from flowmachine.core.cache import reset_cache
-
-
-@pytest.fixture(autouse=True)
-def flowmachine_connect():
-    con = flowmachine.connect()
-    yield con
-    reset_cache(con)
-    con.engine.dispose()  # Close the connection
-    Query.redis.flushdb()  # Empty the redis
-    del Query.connection  # Ensure we recreate everything at next use
-
 
 @pytest.fixture
 def get_dataframe(flowmachine_connect):
